# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para o ecossistema Trampaí
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  client: "client",
  provider: "provider",
  admin: "admin",
} as const;

export type UserVerificationStatus =
  (typeof UserVerificationStatus)[keyof typeof UserVerificationStatus];

export const UserVerificationStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface User {
  id: string;
  email: string;
  name: string;
  role: UserRole;
  avatarUrl?: string | null;
  creditBalance: number;
  verificationStatus: UserVerificationStatus;
  city?: string | null;
  neighborhood?: string | null;
  state?: string | null;
  phone?: string | null;
  referralCode?: string | null;
}

export type UserUpdateRole =
  (typeof UserUpdateRole)[keyof typeof UserUpdateRole];

export const UserUpdateRole = {
  client: "client",
  provider: "provider",
} as const;

export type UserUpdateVerificationStatus =
  (typeof UserUpdateVerificationStatus)[keyof typeof UserUpdateVerificationStatus];

export const UserUpdateVerificationStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface UserUpdate {
  name?: string;
  phone?: string;
  city?: string;
  neighborhood?: string;
  state?: string;
  acceptedTerms?: boolean;
  role?: UserUpdateRole;
  verificationStatus?: UserUpdateVerificationStatus;
  referredById?: string | null;
}

export type UserCreateRole =
  (typeof UserCreateRole)[keyof typeof UserCreateRole];

export const UserCreateRole = {
  client: "client",
  provider: "provider",
} as const;

export interface UserCreate {
  email: string;
  name: string;
  password: string;
  role: UserCreateRole;
}

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  open: "open",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Category {
  id: string;
  name: string;
  icon: string;
}

export interface Job {
  id: string;
  title: string;
  description: string;
  categoryId: string;
  clientId: string;
  budget?: number | null;
  status: JobStatus;
  location: string;
  createdAt: string;
  category?: Category;
  client?: User;
  images?: string[] | null;
}

export interface JobCreate {
  title: string;
  description: string;
  categoryId: string;
  budget?: number | null;
  location: string;
  images?: string[] | null;
}

export type LoginBody = {
  email: string;
  password: string;
};

export type Login200 = {
  token?: string;
  user?: User;
};

export type UnlockJobBodyType =
  (typeof UnlockJobBodyType)[keyof typeof UnlockJobBodyType];

export const UnlockJobBodyType = {
  NORMAL: "NORMAL",
  PLUS: "PLUS",
  EXCLUSIVE: "EXCLUSIVE",
} as const;

export type UnlockJobBody = {
  type: UnlockJobBodyType;
};

export type UnlockJob200 = {
  whatsappLink?: string;
  job?: Job;
};