# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Trampai API
  version: 0.1.0
  description: API para o ecossistema Trampaí
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Operações de saúde do sistema
  - name: auth
    description: Autenticação de usuários
  - name: jobs
    description: Gerenciamento de trampos
  - name: categories
    description: Categorias de serviços

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Saudável
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Registra um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Autentica um usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email, password]
      responses:
        "200":
          description: Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: "#/components/schemas/User" }

  /auth/me:
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Atualiza os dados do usuário logado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: Lista todos os trampos abertos
      responses:
        "200":
          description: Lista de jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Cria um novo trampo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreate"
      responses:
        "201":
          description: Job criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{id}/unlock:
    post:
      operationId: unlockJob
      tags: [jobs]
      summary: Desbloqueia o contato de um trampo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string, enum: [NORMAL, PLUS, EXCLUSIVE] }
              required: [type]
      responses:
        "200":
          description: Lead desbloqueado
          content:
            application/json:
              schema:
                type: object
                properties:
                  whatsappLink: { type: string }
                  job: { $ref: "#/components/schemas/Job" }

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: Lista categorias de serviços
      responses:
        "200":
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]

    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string, enum: [client, provider, admin] }
        avatarUrl: { type: string, nullable: true }
        creditBalance: { type: integer, default: 0 }
        verificationStatus: { type: string, enum: [PENDING, APPROVED, REJECTED], default: PENDING }
        city: { type: string, nullable: true }
        neighborhood: { type: string, nullable: true }
        state: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        referralCode: { type: string, nullable: true }
      required: [id, email, name, role, creditBalance, verificationStatus]

    UserUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        city: { type: string }
        neighborhood: { type: string }
        state: { type: string }
        acceptedTerms: { type: boolean }
        role: { type: string, enum: [client, provider] }
        verificationStatus: { type: string, enum: [PENDING, APPROVED, REJECTED] }
        referredById: { type: string, nullable: true }

    UserCreate:
      type: object
      properties:
        email: { type: string, format: email }
        name: { type: string }
        password: { type: string }
        role: { type: string, enum: [client, provider] }
      required: [email, name, password, role]

    Job:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        categoryId: { type: string, format: uuid }
        clientId: { type: string, format: uuid }
        budget: { type: integer, nullable: true }
        status: { type: string, enum: [open, in_progress, completed, cancelled] }
        location: { type: string }
        createdAt: { type: string, format: date-time }
        category: { $ref: "#/components/schemas/Category" }
        client: { $ref: "#/components/schemas/User" }
        images:
          type: array
          items:
            type: string
          nullable: true
      required: [id, title, description, categoryId, clientId, status, location, createdAt]

    JobCreate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        categoryId: { type: string, format: uuid }
        budget: { type: integer, nullable: true }
        location: { type: string }
        images:
          type: array
          items:
            type: string
          nullable: true
      required: [title, description, categoryId, location]

    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        icon: { type: string }
      required: [id, name, icon]