# packages/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para o ecossistema Trampaí
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";
import type { JobStatus } from "./jobStatus";
import type { User } from "./user";

export interface Job {
  id: string;
  title: string;
  description: string;
  categoryId: string;
  clientId: string;
  budget?: number | null;
  status: JobStatus;
  location: string;
  createdAt: Date;
  category?: Category;
  client?: User;
  images?: string[] | null;
}