# packages/api-zod/src/generated/types/jobCreate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para o ecossistema Trampaí
 * OpenAPI spec version: 0.1.0
 */

export interface JobCreate {
  title: string;
  description: string;
  categoryId: string;
  budget?: number | null;
  location: string;
  images?: string[] | null;
}